+++ src/ServerFactory.php
@@ -3,10 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Server\Control;
 
-use Innmind\Server\Control\{
-    Servers\Unix,
-    Exception\UnsupportedOperatingSystem,
-};
+use Innmind\Server\Control\Exception\UnsupportedOperatingSystem;
 use Innmind\TimeContinuum\{
     Clock,
     Period,
@@ -29,7 +26,7 @@ final class ServerFactory
         switch (\PHP_OS) {
             case 'Darwin':
             case 'Linux':
-                return Unix::of($clock, $io, $halt, $grace);
+                return Server::new($clock, $io, $halt, $grace);
         }
 
         throw new UnsupportedOperatingSystem;