+++ src/Servers/Implementation.php
@@ -1,0 +1,21 @@
+<?php
+declare(strict_types = 1);
+
+namespace Innmind\Server\Control\Servers;
+
+use Innmind\Server\Control\Server\{
+    Processes,
+    Volumes,
+};
+
+/**
+ * @internal
+ */
+interface Implementation
+{
+    #[\NoDiscard]
+    public function processes(): Processes;
+
+    #[\NoDiscard]
+    public function volumes(): Volumes;
+}